+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 ## [v2.3.6.dev0]
 
 ### Added
-- 
+- [Storage] Added MinIO storage backend
+- [Core] Allow to pass function args as part of the invocation payload in FaaS backends
 
 ### Changed
 - [AWS Lambda] Use layer from Klayers API for pre-compiled Amazon Linux numpy binaries
@@ -15,6 +16,7 @@
 - [Core] Fixed an unusual inconsistency in configuration between 'backend' and 'mode' parameters
 - [Core] Fixed customized_runtime feature
 - [Core] Fixed get_result() execution after wait() when throw_except is set to False
+- [Core] Fixed internal executions
 
 
 ## [v2.3.5]